requirements.txt
args==0.1.0
clint==0.4.1
Django==1.7.4
django-ajax-selects==1.3.5
django-extensions==1.5.0
django-grappelli==2.6.3
django-model-utils==2.2
django-registration-redux==1.1
django-shell-plus==1.1.5
django-social-auth==0.7.28
gnureadline==6.3.3
httplib2==0.9
ipdb==0.8
ipython==2.4.1
oauth2==1.5.211
python-openid==2.2.5
requests==2.5.1
six==1.9.0
